# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Monday API Automation",
    "description": "API para automatizar a criação de estrutura de pastas baseada em demandas do Monday.com, definindo analistas responsáveis por produto.",
    "version": "1.0.0",
    "contact": {
      "name": "Nathan Silva - Fagron Tech",
      "email": "nathan.silva@fagrontech.com.br"
    },
    "license": {
      "name": "ISC",
      "url": "https://opensource.org/licenses/ISC"
    }
  },
  "servers": [
    {
      "url": "https://monday-api-automation.onrender.com",
      "description": "Servidor de Produção (Render) - GitHub: https://github.com/Nathan-Paranhos/monday-api-automation"
    },
    {
      "url": "http://localhost:3000",
      "description": "Servidor de Desenvolvimento Local"
    }
  ],
  "paths": {
...
    "/webhook/history": {
      "get": {
        "tags": ["Webhook"],
        "summary": "Histórico de Webhooks",
        "description": "Recupera o histórico de webhooks processados",
        "operationId": "webhookHistory",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {"type": "integer", "default": 1}
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {"type": "integer", "default": 10}
          }
        ],
        "responses": {
          "200": {
            "description": "Histórico de webhooks",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "webhooks": {"type": "array"},
                    "total": {"type": "integer"},
                    "page": {"type": "integer"},
                    "limit": {"type": "integer"}
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/webhook/monday": {
      "post": {
        "tags": ["Webhook"],
        "summary": "Webhook do Monday.com",
        "description": "Endpoint para receber notificações automáticas do Monday.com quando o produto 'BOT' é selecionado",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "event": {
                    "type": "object",
                    "properties": {
                      "type": {
                        "type": "string",
                        "example": "change_column_value"
                      },
                      "columnId": {
                        "type": "string",
                        "example": "produto"
                      },
                      "pulseId": {
                        "type": "string",
                        "example": "123456789"
                      }
                    }
                  }
                },
                "required": ["event"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Webhook processado com sucesso",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "example": "Webhook recebido com sucesso"
                }
              }
            }
          },
          "500": {
            "description": "Erro ao processar webhook",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "example": "Erro ao processar webhook"
                }
              }
            }
          }
        }
      }
    }
  },
...
}